Added Validations on Supplier,Category,Product.  Added Global Exception Handling Middleware

+++ Ecommerce.API/Service/SupplierService.cs
@@ -1,4 +1,5 @@
 ﻿using Ecommerce.API.Entities;
+using Ecommerce.API.Exceptions;
 using Ecommerce.API.Models.SupplierDtos;
 using Ecommerce.API.Repository.Interfaces;
 using Ecommerce.API.Service.Interfaces;
@@ -10,6 +11,10 @@ namespace Ecommerce.API.Service
     {
         public async Task<int> CreateNewSupplierAsync(SupplierForCreatingDto model)
         {
+            if (model is null) throw new BadRequestException("Request model required!");
+
+            if (model.SupplierName.Length > 100) throw new BadRequestException("Product Name Length Can't Exceed 100!");
+
             var newSupplier = mapper.Map<Supplier>(model);
             await supplierRepository.AddAsync(newSupplier);
             return await supplierRepository.SaveAsync();
@@ -19,24 +24,28 @@ namespace Ecommerce.API.Service
         {
             var suppliers = (await supplierRepository.GetAllAsync()).Items;
 
-            // MAPPER  DIDNT WORK
-            //var mapped = mapper.Map<IEnumerable<SupplierForGettingDto>>(suppliers);
+            
+            var mapped = mapper.Map<IEnumerable<SupplierForGettingDto>>(suppliers);
 
-            //return mapped;
-            ////MANUAL MAPPING WORKED
-            return suppliers.Select(c => new SupplierForGettingDto()
-            {
-                Id = c.Id,
-                SuplierName = c.SupplierName
-            });
+            return mapped;
+            ////MANUAL MAPPING
+            //return suppliers.Select(c => new SupplierForGettingDto()
+            //{
+            //    Id = c.Id,
+            //    SuplierName = c.SupplierName
+            //});
         }
 
         public async Task<int> UpdateSupplierAsync(SupplierForUpdatingDto model)
         {
+            if (model is null) throw new BadRequestException("Request Model Required");
+
+            if (model.Id == Guid.Empty) throw new BadRequestException("Supplier Id Required!");
+
             var supplier = await supplierRepository.GetAsync(p => p.Id == model.Id);
 
             if (supplier is null)
-                return 0;
+                throw new NotFoundException($"Supplier With Id {model.Id} Not Found!");
 
             supplier.SupplierName = model.SupplierName;
 
@@ -45,10 +54,11 @@ namespace Ecommerce.API.Service
 
         public async Task<int> DeleteSupplierAsync(Guid Id)
         {
+            if (Id == Guid.Empty) throw new BadRequestException("Supplier Id Required");
             var supplier = await supplierRepository.GetAsync(p => p.Id == Id);
 
             if (supplier is null)
-                return 0;
+                throw new NotFoundException($"Supplier With Id {Id} Not Found!");
 
             supplierRepository.Remove(supplier);
            return await supplierRepository.SaveAsync();
